refactor(aq25): split daemon launch and bare cli composition
